Treats - Added Treats that are spawned when a Brick is destroyed. They launch in a direction and fall with gravity. - PlayState has a vector for Treats, deleting them when they go offscreen. - PlayState returns to level select automatically when there are no stages left in the current stage. - Can now get the current State from Game. Makes it easier to add stuff in from GameObjects. - Made Add and Remove in State virtual so deriving states can do special things according to type. Used in PlayState for the Treat vector. - Added nullptr checks to Add and Remove in State.

+++ src/Game/LevelSelectState.cpp
@@ -80,7 +80,7 @@ void LevelSelectState::Init()
 	int i = 0;
 	for (auto it = levelDatas.begin(); it != levelDatas.end(); it++)
 	{
-		Button* levelButton = new Button(sf::Vector2f((float)stageSize.x / 2.0f, i * 60.0f + 30.0f), sf::Vector2f(stageSize.x, 60.0f), (ButtonStringCallback)std::bind(&LevelSelectState::LevelButtonPressed, this, _1));
+		Button* levelButton = new Button(sf::Vector2f((float)stageSize.x / 2.0f, i * 60.0f + 30.0f), sf::Vector2f((float)stageSize.x, 60.0f), (ButtonStringCallback)std::bind(&LevelSelectState::LevelButtonPressed, this, _1));
 		levelButton->LoadButtonGraphic(G::GetAssetManager()->GetTexture("button.png"), sf::Vector2f(18, 18), sf::Vector2f(6, 6));
 		levelButton->GetText()->setFont(*G::GetAssetManager()->GetFont("standur.ttf"));
 		levelButton->GetText()->setFillColor(sf::Color::Black);
